requirements.txt
colorama==0.4.6
numpy==1.26.0
PrettyPrintTree==1.36.3
